style: Create file_new.go
